requirements.txt
polygon-api-client
pymongo
certifi
alpaca-py
yfinance
scipy
pandas
cython
fastapi
uvicorn
motor
python-dotenv
stats
PyWavelets
ripser
datetime
ta-lib-python
pandas-ta